backend/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
pydantic
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
sendgrid